Clean code and fix crash with early use of tprintf

+++ src/arch/intsimdmatrixsse.cpp
@@ -2,7 +2,6 @@
 // File:        intsindmatrixsse.cpp
 // Description: SSE implementation of 8-bit int SIMD matrix multiply.
 // Author:      Ray Smith
-// Created:     Tue Aug 23 13:58:49 PST 2017
 //
 // (C) Copyright 2017, Google Inc.
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -19,7 +18,6 @@
 #include "intsimdmatrixsse.h"
 
 #include <cstdint>
-#include <vector>
 #include "dotproductsse.h"
 
 namespace tesseract {
@@ -29,9 +27,9 @@ namespace tesseract {
 static void PartialMatrixDotVector1(const int8_t* wi, const double* scales,
                                     const int8_t* u, int num_in, int num_out,
                                     double* v) {
-  int total = IntDotProductSSE(u, wi, num_in);
+  double total = IntDotProductSSE(u, wi, num_in);
   // Add in the bias and correct for integer values.
-  *v = (static_cast<double>(total) / INT8_MAX + wi[num_in]) * *scales;
+  *v = (total / INT8_MAX + wi[num_in]) * *scales;
 }
 #endif  // __SSE4_1__
